+++ doc/source/changes.rst
@@ -9,7 +9,8 @@ New in 0.21
 
 * :class:`MneExperiment`:
 
-  - New epoch parameter: ``trigger_shift`` (see :attr:`MneExperiment.epochs`).
+  - New epoch parameters: ``trigger_shift`` and ``vars`` (see
+    :attr:`MneExperiment.epochs`).
 
 * New :class:`Factor` method:  :meth:`Factor.floodfill`.